Split up instant launch code into multiple files
